requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiokafka==0.13.0
    # via faststream
annotated-types==0.7.0
    # via pydantic
anthropic==0.84.0
    # via calfkit
anyio==4.12.1
    # via
    #   anthropic
    #   fast-depends
    #   faststream
    #   httpx
    #   openai
async-timeout==5.0.1
    # via aiokafka
calfkit==0.1.10
    # via crypto-daytrading-arena (pyproject.toml)
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
colorama==0.4.6
    # via griffecli
distro==1.9.0
    # via
    #   anthropic
    #   openai
docstring-parser==0.17.0
    # via anthropic
fast-depends==3.0.5
    # via faststream
faststream==0.6.6
    # via calfkit
genai-prices==0.0.54
    # via calfkit
griffe==2.0.0
    # via calfkit
griffecli==2.0.0
    # via griffe
griffelib==2.0.0
    # via
    #   griffe
    #   griffecli
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   crypto-daytrading-arena (pyproject.toml)
    #   anthropic
    #   calfkit
    #   genai-prices
    #   openai
    #   pydantic-graph
idna==3.11
    # via
    #   anyio
    #   httpx
importlib-metadata==8.7.1
    # via opentelemetry-api
jiter==0.13.0
    # via
    #   anthropic
    #   openai
logfire-api==4.25.0
    # via pydantic-graph
markdown-it-py==4.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
openai==2.24.0
    # via calfkit
opentelemetry-api==1.39.1
    # via calfkit
packaging==26.0
    # via aiokafka
plotext==5.3.2
    # via crypto-daytrading-arena (pyproject.toml)
pydantic==2.12.5
    # via
    #   crypto-daytrading-arena (pyproject.toml)
    #   anthropic
    #   calfkit
    #   fast-depends
    #   genai-prices
    #   openai
    #   pydantic-graph
pydantic-core==2.41.5
    # via pydantic
pydantic-graph==1.63.0
    # via calfkit
pygments==2.19.2
    # via rich
python-dotenv==1.2.1
    # via
    #   crypto-daytrading-arena (pyproject.toml)
    #   calfkit
rich==14.3.3
    # via crypto-daytrading-arena (pyproject.toml)
sniffio==1.3.1
    # via
    #   anthropic
    #   openai
sympy==1.14.0
    # via crypto-daytrading-arena (pyproject.toml)
tqdm==4.67.3
    # via openai
typing-extensions==4.15.0
    # via
    #   aiokafka
    #   anthropic
    #   anyio
    #   fast-depends
    #   faststream
    #   openai
    #   opentelemetry-api
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   calfkit
    #   pydantic
    #   pydantic-graph
uuid-utils==0.14.1
    # via calfkit
websockets==16.0
    # via crypto-daytrading-arena (pyproject.toml)
zipp==3.23.0
    # via importlib-metadata
